pyproject.toml
[tool.poetry]
name = "shop-backend"
version = "0.1.0"
description = ""
authors = ["Nikolai Zinchenko <nikolaizinchenko94@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
djangorestframework = "^3.14.0"
python-dotenv = "^1.0.0"
environs = "^10.0.0"
psycopg2-binary = "^2.9.9"
django-filter = "^23.5"
django-filer = "^3.1.1"
requests = "^2.31.0"
celery = "^5.3.6"
redis = "^5.0.1"
django-nested-inline = "^0.4.6"


[tool.poetry.group.dev.dependencies]
flake8 = "^6.1.0"
black = "^23.12.1"
isort = "^5.13.2"
coverage = "^7.3.4"
mypy = "^1.8.0"
django-stubs = "^4.2.7"
types-requests = "^2.31.0.10"
types-pillow = "^10.1.0.2"
celery-stubs = "^0.1.3"
djangorestframework-stubs = "^3.14.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.coverage.run]
omit = [
    "*/migrations/*",
    "*/tests/*",
    "__init__.py",
    ]

[tool.mypy]
plugins = ["mypy_django_plugin.main", "mypy_drf_plugin.main"]
exclude = ["migrations", "adaptive_images"]
follow_imports = 'silent'

[tool.django-stubs]
django_settings_module = "core.settings"

[flake8]
max-line-length = 99
max-doc-length = 79

[tool.black]
line-length = 99
